+++ internal/envirements/envirements.go
@@ -6,18 +6,18 @@ import (
 )
 
 type EnvConfig struct {
-	EnvRunAddr   string
-	EnvResultURL string
+	RunAddr   string
+	ResultURL string
 }
 
 // ParseEnv Get system environments
 func ParseEnv() *EnvConfig {
 	env := &EnvConfig{}
-	env.EnvRunAddr = os.Getenv("RUN_ADDR")
-	env.EnvResultURL = os.Getenv("SERVER_ADDRESS")
+	env.RunAddr = os.Getenv("RUN_ADDR")
+	env.ResultURL = os.Getenv("SERVER_ADDRESS")
 	// парсим переданные серверу аргументы в зарегистрированные переменные
-	if !strings.HasPrefix(env.EnvResultURL, "http://") && env.EnvResultURL != "" {
-		env.EnvResultURL = "http://" + env.EnvResultURL
+	if !strings.HasPrefix(env.ResultURL, "http://") && env.ResultURL != "" {
+		env.ResultURL = "http://" + env.ResultURL
 	}
 	return env
 }
