--- internal/logger/logger.go
@@ -1,1 +1,0 @@
-package logger
